+++ src/vec4.rs
@@ -3,7 +3,7 @@ use std::{
     fmt::Display,
     ops::{
         Add, AddAssign, Div, DivAssign, Index, IndexMut, Mul, MulAssign, Neg,
-        Rem, RemAssign, Sub, SubAssign,
+        Not, Rem, RemAssign, Sub, SubAssign,
     },
 };
 
@@ -254,6 +254,17 @@ impl<T: Neg> Neg for Vec4<T> {
 
     fn neg(self) -> Self::Output {
         self.map(|x| -x)
+    }
+}
+
+impl<T> Not for Vec4<T>
+where
+    T: Not,
+{
+    type Output = Vec4<T::Output>;
+
+    fn not(self) -> Self::Output {
+        self.map(|a| !a)
     }
 }