+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### New features
 - Add methods to `RectExt`: `moved_by`, `set_right`, `set_bottom`, `set_left`,
   `set_top`.
+- Implement `Not` for `Vec2`, `Vec3` and `Vec4`.
 
 ## v0.4.0
 ### Breaking changes